Fix the native test environment picking up board = esp32dev
[env] is PlatformIO's global section, inherited by every environment including
native, so `board = esp32dev` reached the host test build and every suite
errored with "Unknown board ID 'esp32dev'" before compiling anything.

Board settings move to a plain [esp32] section, which is only applied to the
environments that name it via `extends`. [env] keeps just the include paths, the
warning flags and the SAW_EXTRA_FLAGS commissioning hook — the things that are
true for a host build and a board build alike.

The four ESP32 environments were unaffected and already built green at 882 KiB,
comfortably inside the 1.5 MB OTA slot.

Also drop two signed/unsigned comparisons in saw_store.cpp: fs::File::read
returns size_t, so comparing against (int)sizeof(...) warned under -Wextra.

+++ firmware/src/saw_store.cpp
@@ -54,7 +54,7 @@ static bool meta_read(LogMeta *m)
     if (!f) {
         return false;
     }
-    bool ok = (f.read((uint8_t *)m, sizeof(*m)) == (int)sizeof(*m));
+    bool ok = (f.read((uint8_t *)m, sizeof(*m)) == sizeof(*m));
     f.close();
     if (!ok || m->magic != META_MAGIC) {
         return false;
@@ -267,7 +267,7 @@ uint32_t saw_store_stream_samples(SawSampleVisitor cb, void *ctx)
             continue;
         }
         SawSampleRecord rec;
-        while (f.read((uint8_t *)&rec, sizeof(rec)) == (int)sizeof(rec)) {
+        while (f.read((uint8_t *)&rec, sizeof(rec)) == sizeof(rec)) {
             uint16_t want = saw_crc16((const uint8_t *)&rec,
                                       sizeof(rec) - sizeof(uint16_t));
             if (want != rec.crc) {
